+++ themes/theme-rui/src/components/Underlay.styles.ts
@@ -3,7 +3,7 @@ import { ThemeComponent, cva } from '@marigold/system';
 export const Underlay: ThemeComponent<'Underlay'> = cva('', {
   variants: {
     variant: {
-      modal: ' bg-black/25 backdrop-blur-sm',
+      modal: ' bg-black/80 backdrop-blur-sm',
     },
   },
 });